move metrics to internal/pkg
