+++ internal/pkg/metrics/query/examples/examples.yaml
