+++ internal/pkg/metrics/query/examples/main.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/erda-project/erda-infra/base/logs"
 	"github.com/erda-project/erda-infra/base/servicehub"
-	"github.com/erda-project/erda/providers/metrics/query"
+	"github.com/erda-project/erda/internal/pkg/metrics/query"
 )
 
 type define struct{}
