+++ internal/pkg/metrics/report/compress.go
