+++ internal/pkg/metrics/report/examples/examples.yaml
