+++ internal/pkg/metrics/report/examples/main.go
@@ -21,7 +21,7 @@ import (
 	"github.com/recallsong/go-utils/logs"
 
 	"github.com/erda-project/erda-infra/base/servicehub"
-	"github.com/erda-project/erda/providers/metrics/report"
+	"github.com/erda-project/erda/internal/pkg/metrics/report"
 )
 
 type define struct{}
