+++ internal/pkg/metrics/report/global.go
