+++ internal/tools/monitor/extensions/loghub/sls-log-to-metric/provider.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/erda-project/erda-infra/base/logs"
 	"github.com/erda-project/erda-infra/base/servicehub"
-	"github.com/erda-project/erda/providers/metrics/report"
+	"github.com/erda-project/erda/internal/pkg/metrics/report"
 )
 
 type (
