+++ internal/tools/pipeline/metrics/pipeline.go
@@ -35,8 +35,8 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/erda-project/erda/apistructs"
+	"github.com/erda-project/erda/internal/pkg/metrics/report"
 	"github.com/erda-project/erda/internal/tools/pipeline/spec"
-	"github.com/erda-project/erda/providers/metrics/report"
 )
 
 const (
