+++ internal/tools/pipeline/metrics/task.go
@@ -21,8 +21,8 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/erda-project/erda/apistructs"
+	"github.com/erda-project/erda/internal/pkg/metrics/report"
 	spec2 "github.com/erda-project/erda/internal/tools/pipeline/spec"
-	"github.com/erda-project/erda/providers/metrics/report"
 )
 
 const (
